+++ Frends.MicrosoftSQL.BulkInsert/Frends.MicrosoftSQL.BulkInsert/Definitions/ColumnMapping.cs
@@ -1,0 +1,22 @@
+﻿namespace Frends.MicrosoftSQL.BulkInsert.Definitions;
+
+/// <summary>
+/// SqlTransactionIsolationLevels
+/// </summary>
+public enum ColumnMapping
+{
+    /// <summary>
+    /// Column mapping is disabled and the bulk insert will insert the data based on the order of the properties in input JSON.
+    /// </summary>
+    JsonPropertyOrder,
+
+    /// <summary>
+    /// Input JSON property names will be used with bulk insert to create column mapping. Column mapping is case sensitive.
+    /// </summary>
+    JsonPropertyNames,
+
+    /// <summary>
+    /// Manual column mapping JSON will be used with bulk insert to create column mapping. Column mapping is case sensitive.
+    /// </summary>
+    ManualColumnMapping,
+}